refactor: extract duplicate stripHopByHop into shared utility
